feat: Add Character Select on Login flow
